+++ src/main/java/nextstep/auth/AuthConfig.java
@@ -1,7 +1,10 @@
 package nextstep.auth;
 
 import lombok.RequiredArgsConstructor;
-import nextstep.auth.authentication.*;
+import nextstep.auth.authentication.filter.chain.BasicAuthenticationFilter;
+import nextstep.auth.authentication.filter.chain.BearerTokenAuthenticationFilter;
+import nextstep.auth.authentication.filter.nonchain.TokenAuthenticationFilter;
+import nextstep.auth.authentication.filter.nonchain.UsernamePasswordAuthenticationFilter;
 import nextstep.auth.authentication.provider.ProviderManager;
 import nextstep.auth.authorization.AuthenticationPrincipalArgumentResolver;
 import nextstep.auth.context.SecurityContextPersistenceFilter;
