+++ src/main/java/nextstep/auth/authentication/filter/chain/AuthenticationChainFilter.java
@@ -1,25 +1,31 @@
-package nextstep.auth.authentication;
+package nextstep.auth.authentication.filter.chain;
 
+import nextstep.auth.authentication.AuthenticationToken;
 import nextstep.auth.context.Authentication;
+import nextstep.auth.context.SecurityContextHolder;
 import org.springframework.web.servlet.HandlerInterceptor;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
+import java.util.Objects;
 
-public abstract class AuthenticationNonChainFilter implements HandlerInterceptor {
+public abstract class AuthenticationChainFilter implements HandlerInterceptor {
 
     @Override
     public boolean preHandle(HttpServletRequest request, HttpServletResponse response, Object handler) throws Exception {
         AuthenticationToken token = createToken(request);
-        Authentication authenticate = authenticate(token);
+        if (Objects.isNull(token)) {
+            return true;
+        }
 
-        return send(authenticate, response);
+        Authentication authentication = authenticate(token);
+        SecurityContextHolder.getContext().setAuthentication(authentication);
+        return true;
     }
 
     protected abstract AuthenticationToken createToken(HttpServletRequest request) throws IOException;
 
     protected abstract Authentication authenticate(AuthenticationToken token);
 
-    protected abstract boolean send(Authentication authentication, HttpServletResponse response) throws IOException;
 }
