+++ src/main/java/nextstep/auth/authentication/filter/chain/BearerTokenAuthenticationFilter.java
@@ -1,31 +1,27 @@
-package nextstep.auth.authentication;
+package nextstep.auth.authentication.filter.chain;
 
 import lombok.RequiredArgsConstructor;
+import nextstep.auth.authentication.AuthenticationToken;
+import nextstep.auth.authentication.AuthorizationExtractor;
+import nextstep.auth.authentication.AuthorizationType;
 import nextstep.auth.authentication.provider.AuthenticationProvider;
 import nextstep.auth.authentication.provider.ProviderManager;
 import nextstep.auth.authentication.provider.ProviderType;
 import nextstep.auth.context.Authentication;
-import org.apache.tomcat.util.codec.binary.Base64;
 
 import javax.servlet.http.HttpServletRequest;
 import java.io.IOException;
 
 @RequiredArgsConstructor
-public class BasicAuthenticationFilter extends AuthenticationChainFilter {
+public class BearerTokenAuthenticationFilter extends AuthenticationChainFilter {
 
     private final ProviderManager providerManager;
 
     @Override
     protected AuthenticationToken createToken(HttpServletRequest request) throws IOException {
         try {
-            String authCredentials = AuthorizationExtractor.extract(request, AuthorizationType.BASIC);
-            String authHeader = new String(Base64.decodeBase64(authCredentials));
-
-            String[] splits = authHeader.split(":");
-            String principal = splits[0];
-            String credentials = splits[1];
-
-            return new AuthenticationToken(principal, credentials);
+            String authCredentials = AuthorizationExtractor.extract(request, AuthorizationType.BEARER);
+            return new AuthenticationToken(authCredentials, authCredentials);
         } catch (Exception e) {
             return null;
         }
@@ -33,7 +29,8 @@ public class BasicAuthenticationFilter extends AuthenticationChainFilter {
 
     @Override
     protected Authentication authenticate(AuthenticationToken token) {
-        AuthenticationProvider authenticationProvider = providerManager.getAuthenticationProvider(ProviderType.USER_PASSWORD);
+        AuthenticationProvider authenticationProvider = providerManager.getAuthenticationProvider(ProviderType.JWT_TOKEN);
         return authenticationProvider.authenticate(token);
     }
+
 }
